+++ src/mesido/esdl/asset_to_component_base.py
@@ -192,7 +192,7 @@ class _AssetToComponentBase:
     primary_port_name_convention = "primary"
     secondary_port_name_convention = "secondary"
 
-    __power_keys = ["power", "maxDischargeRate", "maxChargeRate"]
+    __power_keys = ["power", "maxDischargeRate", "maxChargeRate", "capacity"]
 
     def __init__(self, **kwargs):
         """
@@ -516,23 +516,43 @@ class _AssetToComponentBase:
             i_max_out = (
                 self._port_to_i_max.get(connected_port, None)
                 if self._port_to_i_max.get(connected_port, False)
-                else max([asset.attributes.get(key, -1) for key in self.__power_keys])
+                else max(
+                    [
+                        asset.attributes.get(key, -1) / connected_port.carrier.voltage
+                        for key in self.__power_keys
+                    ]
+                )
             )
             i_nom_out = (
                 self._port_to_i_nominal.get(connected_port, None)
                 if self._port_to_i_nominal.get(connected_port, False)
-                else max([asset.attributes.get(key, -1) for key in self.__power_keys])
+                else max(
+                    [
+                        asset.attributes.get(key, -1) / connected_port.carrier.voltage / 2.0
+                        for key in self.__power_keys
+                    ]
+                )
             )
             connected_port = asset.in_ports[0].connectedTo[0]
             i_max_in = (
                 self._port_to_i_max.get(connected_port, None)
                 if self._port_to_i_max.get(connected_port, False)
-                else max([asset.attributes.get(key, -1) for key in self.__power_keys])
+                else max(
+                    [
+                        asset.attributes.get(key, -1) / connected_port.carrier.voltage
+                        for key in self.__power_keys
+                    ]
+                )
             )
             i_nom_in = (
                 self._port_to_i_nominal.get(connected_port, None)
                 if self._port_to_i_nominal.get(connected_port, False)
-                else max([asset.attributes.get(key, -1) for key in self.__power_keys])
+                else max(
+                    [
+                        asset.attributes.get(key, -1) / connected_port.carrier.voltage / 2.0
+                        for key in self.__power_keys
+                    ]
+                )
             )
             if i_nom_in > 0.0 and i_nom_out > 0.0:
                 self._port_to_i_nominal[asset.in_ports[0]] = i_nom_in
@@ -551,12 +571,22 @@ class _AssetToComponentBase:
             i_max = (
                 self._port_to_i_max.get(connected_port, None)
                 if self._port_to_i_max.get(connected_port, False)
-                else max([asset.attributes.get(key, -1) for key in self.__power_keys])
+                else max(
+                    [
+                        asset.attributes.get(key, -1) / connected_port.carrier.voltage
+                        for key in self.__power_keys
+                    ]
+                )
             )
             i_nom = (
                 self._port_to_i_nominal.get(connected_port, None)
                 if self._port_to_i_nominal.get(connected_port, False)
-                else max([asset.attributes.get(key, -1) for key in self.__power_keys])
+                else max(
+                    [
+                        asset.attributes.get(key, -1) / connected_port.carrier.voltage / 2.0
+                        for key in self.__power_keys
+                    ]
+                )
             )
             if i_max > 0.0:
                 self._set_electricity_current_nominal_and_max(asset, i_max, i_nom)
@@ -578,12 +608,22 @@ class _AssetToComponentBase:
                     i_max = (
                         self._port_to_i_max.get(connected_port, None)
                         if self._port_to_i_max.get(connected_port, False)
-                        else max([asset.attributes.get(key, -1) for key in self.__power_keys])
+                        else max(
+                            [
+                                asset.attributes.get(key, -1) / connected_port.carrier.voltage
+                                for key in self.__power_keys
+                            ]
+                        )
                     )
                     i_nom = (
                         self._port_to_i_nominal.get(connected_port, None)
                         if self._port_to_i_nominal.get(connected_port, False)
-                        else max([asset.attributes.get(key, -1) for key in self.__power_keys])
+                        else max(
+                            [
+                                asset.attributes.get(key, -1) / connected_port.carrier.voltage / 2.0
+                                for key in self.__power_keys
+                            ]
+                        )
                     )
                     if i_max > 0.0:
                         self._set_electricity_current_nominal_and_max(asset, i_max, i_nom)