+++ src/mesido/gas_physics_mixin.py
@@ -142,6 +142,10 @@ class GasPhysicsMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPr
         self.__gas_storage_discharge_nominals = {}
         self.__gas_storage_discharge_map = {}
 
+        # Map for setting port variable nominals in the case they were not set during the model
+        # parsing (logical links).
+        self.__gas_node_variable_nominal = {}
+
     def gas_carriers(self):
         """
         This function should be overwritten by the problem and should give a dict with the
@@ -271,6 +275,37 @@ class GasPhysicsMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPr
                     f"{storage}.GasIn.Q"
                 )
 
+        # Setting the node nominals using the connected assets.
+        for node, connected_assets in self.energy_system_topology.gas_nodes.items():
+            nominals = {}
+            for var in ["Q", "H", "Hydraulic_power"]:
+                nominals[var] = []
+                for _, (asset, _orientation) in connected_assets.items():
+                    var_nom = self.variable_nominal(f"{asset}.GasOut.{var}")
+                    if var_nom != 1:
+                        nominals[var].append(var_nom)
+                    elif self.variable_nominal(f"{asset}.GasIn.{var}") != 1:
+                        nominals[var].append(self.variable_nominal(f"{asset}.GasIn.{var}"))
+                    else:
+                        nominals[var].append(1)
+
+                for i in range(len(connected_assets)):
+                    if self.variable_nominal(f"{node}.GasConn[{i + 1}].{var}") == 1:
+                        if nominals[var][i] != 1:
+                            # Here we set a nominal based directly on the connected asset.
+                            self.__gas_node_variable_nominal[f"{node}.GasConn[{i + 1}].{var}"] = (
+                                nominals[var][i]
+                            )
+                        else:
+                            # Here we set a nominal based on median of all the connected assets to
+                            # the node. This is specifically done when we have a logical link for
+                            # node to node. In this case we cannot set the nominal based on the
+                            # connected node, hence we assume a node has at least one not node
+                            # asset connected to it.
+                            self.__gas_node_variable_nominal[f"{node}.GasConn[{i + 1}].{var}"] = (
+                                np.median([x for x in nominals[var] if x != 1])
+                            )
+
     def energy_system_options(self):
         r"""
         Returns a dictionary of milp network specific options.
@@ -334,6 +369,8 @@ class GasPhysicsMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPr
             return self.__gas_pipe_head_loss_nominals[variable]
         elif variable in self.__gas_storage_discharge_nominals:
             return self.__gas_storage_discharge_nominals[variable]
+        elif variable in self.__gas_node_variable_nominal:
+            return self.__gas_node_variable_nominal[variable]
         else:
             return super().variable_nominal(variable)
 
@@ -596,11 +633,11 @@ class GasPhysicsMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPr
             constraints.extend(
                 self._gn_head_loss_class._pipe_head_loss_path_constraints(self, ensemble_member)
             )
-        constraints.extend(
-            self._gn_head_loss_class._pipe_hydraulic_power_path_constraints(
-                self, self.__maximum_total_head_loss, ensemble_member
+            constraints.extend(
+                self._gn_head_loss_class._pipe_hydraulic_power_path_constraints(
+                    self, self.__maximum_total_head_loss, ensemble_member
+                )
             )
-        )
         constraints.extend(self.__flow_direction_path_constraints(ensemble_member))
         constraints.extend(self.__gas_node_hydraulic_power_mixing_path_constraints(ensemble_member))
         constraints.extend(self.__gas_storage_discharge_path_constraints(ensemble_member))